+++ tensorflow-sys-runtime/src/c_api.rs
@@ -1,3 +1,4 @@
+include!("runtime.rs");
 link! {
 /* automatically generated by rust-bindgen 0.59.2 */
