+++ tensorflow-sys/Cargo.toml
@@ -36,7 +36,6 @@ zip = "0.6.2"
 [features]
 tensorflow_gpu = []
 eager = []
-runtime_linking = ["lazy_static", "libloading", "cfg-if", "log"]
 # This is for testing purposes; users should not use this.
 examples_system_alloc = []
 private-docs-rs = [] # DO NOT RELY ON THIS
