--- tensorflow-sys/src/runtime_linking/mod.rs
@@ -1,4 +1,0 @@
-include!("runtime.rs");
-include!("c_api.rs");
-include!("types.rs");
-include!("finder.rs");
